refactor(runtime): the bound is stated once, and the encoder takes the wider integer
    Four follow-ups to this branch's review, none of them behavioural.

    `spec_digits` rejects rather than answering a capped number a caller compares again, so
    each field's limit is named once. `ZRT_FMT_BODY` is written as the arithmetic its
    comment described, so the compiler checks the reasoning. The int whitelist is a `strchr`
    of the set every other statement of it is a set.

    `zrt_utf8_len`/`zrt_utf8_encode` take an `int64_t`, like `zrt_is_rune`: a caller
    narrowing first would render `{0x100000041:c}` as `A`, a valid rendering of a code point
    nobody asked for. The range answer belongs to the function that owns the range.

+++ src/runtime/csrc/zergrt.h
@@ -698,8 +698,8 @@ int32_t zrt_conv_rune(int64_t v);
  * points a str can hold" is two answers waiting to disagree. zrt_utf8_len answers 0 for a
  * code point no str can hold — a surrogate, out of range, or U+0000, whose byte is a NUL —
  * and zrt_utf8_encode then writes nothing. */
-int zrt_utf8_len(int32_t cp);
-int zrt_utf8_encode(int32_t cp, char *out);
+int zrt_utf8_len(int64_t cp);
+int zrt_utf8_encode(int64_t cp, char *out);
 
 /* `a // b` is floor division whose result is always an `int` (docs/core/types.md). For
  * two integers that IS zrt_div_i64 — the language has one integer division and `//` is
